# [Draft]Use ValueStringBuilder in Encode/DecodeName