--- a/src/libraries/System.Private.Xml/src/System/Xml/XmlConvert.cs
+++ b/src/libraries/System.Private.Xml/src/System/Xml/XmlConvert.cs
@@ -33,6 +33,7 @@ namespace System.Xml
     /// </devdoc>
     public class XmlConvert
     {
+        private const int MaxStackAllocatedChars = 128;
         internal static char[] crt = new char[] { '\n', '\r', '\t' };
 
         /// <devdoc>
@@ -107,7 +108,7 @@ namespace System.Xml
             ValueStringBuilder bufBld =
                 matchPos == -1
                     ? default
-                    : length < 50
+                    : length < MaxStackAllocatedChars
                         ? new ValueStringBuilder(stackalloc char[length])
                         : new ValueStringBuilder(length);
 
@@ -220,7 +221,7 @@ namespace System.Xml
             ValueStringBuilder bufBld =
                 matchPos == -1
                     ? default
-                    : length < 50
+                    : length < MaxStackAllocatedChars
                         ? new ValueStringBuilder(stackalloc char[length])
                         : new ValueStringBuilder(length);