Follow up of [Question](https://github.com/dotnet/runtime/issues/65884)

Results:
Before

|                          Method |         Mean |      Error |     StdDev |       Median |          Min |          Max |  Gen 0 | Allocated |
|-------------------------------- |-------------:|-----------:|-----------:|-------------:|-------------:|-------------:|-------:|----------:|
|                      EncodeName |    81.036 ns |  0.9210 ns |  0.7690 ns |    80.738 ns |    80.248 ns |    82.882 ns | 0.0267 |     224 B |
|               EncodeName_Larger |    93.642 ns |  1.4571 ns |  1.3629 ns |    93.391 ns |    89.993 ns |    95.955 ns |      - |         - |
|              EncodeName_Invalid |   188.668 ns |  1.7830 ns |  1.4889 ns |   189.065 ns |   183.973 ns |   189.956 ns | 0.0523 |     440 B |
|       EncodeName_Invalid_Larger | 1,569.636 ns | 19.6314 ns | 19.2807 ns | 1,564.548 ns | 1,529.487 ns | 1,615.296 ns | 0.4290 |    3608 B |
|                      DecodeName |     5.210 ns |  0.0916 ns |  0.0857 ns |     5.219 ns |     5.042 ns |     5.348 ns |      - |         - |
|               DecodeName_Larger |     6.187 ns |  0.0345 ns |  0.0306 ns |     6.188 ns |     6.145 ns |     6.259 ns |      - |         - |
|              DecodeName_Invalid |   924.028 ns |  4.9796 ns |  4.6579 ns |   924.000 ns |   917.484 ns |   930.695 ns | 0.2408 |    2040 B |
|       DecodeName_Invalid_Larger | 1,048.205 ns |  7.7555 ns |  7.2545 ns | 1,047.681 ns | 1,037.933 ns | 1,061.621 ns | 0.3364 |    2840 B |



After
|                          Method |         Mean |      Error |     StdDev |       Median |          Min |          Max |  Gen 0 | Allocated |
|-------------------------------- |-------------:|-----------:|-----------:|-------------:|-------------:|-------------:|-------:|----------:|
|                      EncodeName |    83.026 ns |  0.1343 ns |  0.1121 ns |    83.042 ns |    82.807 ns |    83.222 ns | 0.0092 |      80 B |
|               EncodeName_Larger |    78.872 ns |  0.3150 ns |  0.2947 ns |    78.913 ns |    78.384 ns |    79.302 ns |      - |         - |
|              EncodeName_Invalid |   145.458 ns |  2.2386 ns |  1.9845 ns |   144.930 ns |   141.253 ns |   148.942 ns | 0.0158 |     136 B |
|       EncodeName_Invalid_Larger | 1,425.675 ns | 15.9157 ns | 13.2903 ns | 1,423.797 ns | 1,411.260 ns | 1,458.079 ns | 0.2784 |    2360 B |
|                      DecodeName |     6.114 ns |  0.1560 ns |  0.1383 ns |     6.104 ns |     5.802 ns |     6.358 ns |      - |         - |
|               DecodeName_Larger |     7.467 ns |  0.0599 ns |  0.0531 ns |     7.449 ns |     7.389 ns |     7.566 ns |      - |         - |
|              DecodeName_Invalid |   908.402 ns |  6.5217 ns |  5.4459 ns |   908.704 ns |   899.138 ns |   916.882 ns | 0.2165 |    1816 B |
|       DecodeName_Invalid_Larger | 1,059.641 ns |  5.5400 ns |  4.9111 ns | 1,059.661 ns | 1,052.079 ns | 1,067.418 ns | 0.2631 |    2216 B |