@TrayanZapryanov 2022-03-03
@krwq , @gfoidl I have two more ideas for improvement here:
1) As regex was used, source generator can be used. But I have no idea how this should be done.
2) Based on your comment that we should optimize only "happy" path, then we can use new method of regex for counting if there are matches in the fast path. If there are then go to slow path like the code is now.

Any thoughts on this?